Tell people they were tagged
Started as "where should the tags page live — dropdown, dashboard or
settings?". Grilling it against adoption turned up two things that made
placement the wrong question.

/tags — the public catalogue, 34 real tags, works logged-out — had ZERO
nav entries anywhere. Chips link to individual tag pages, never to the
list, so browse was reachable only by typing the URL. Meanwhile the page
in the menu was the personal one a normal user needs least.

And the real finding: nothing in this app tells you anything happened.
/alerts is a moderation surface; there is no notification system at all.
So someone could label you in public and you would simply never find
out. Moving a page between menus only changes how hard it is to go
looking — for people who already know tagging exists. The missing step
was the middle of the flywheel: A tags B, B FINDS OUT, B tags C.

Three changes:

1. Browse gets a quiet line under the search box on /. Beside search
   rather than in the nav, because that's the "I don't know what to
   type" half, and the page works logged-out where the dashboard
   doesn't.

2. Your tags moves to Settings as a 4th tab. The tab bar had a
   documented squeeze — three labels over a 339px track at 375px,
   fitting only after a padding fix — so "About & support" becomes
   "About". Measured after: pill 305px in a 375px track, no overflow.
   MyTagsPage's body is extracted to YourTagsPanel so there's one
   implementation; /tags/mine redirects, deliberately NOT behind
   RequireAuth so a logged-out visitor's `next` lands on the canonical
   destination rather than bouncing through a dead URL.

3. TaggedYouModule on the dashboard — the actual adoption work.
   "Tanja tagged you as LFO", her name linked to her profile, which is
   where the tag picker already lives, so the loop closes without a new
   CTA. No "tag them back" button: manufactured reciprocity pollutes the
   data the trust filter depends on. Event, not wallpaper — shows only
   what's new since you last looked, self-hides otherwise. Watermark is
   per-account localStorage following ScoringStatusBar, whose own
   comments record why a global key was a bug.

Needed one service change: ProfileTag had no timestamp, groupByTag
discarded the `at` that NormalizedAssertion already carries. Added
addedAt, mirroring TargetSupport on the tag-page path. Self-assertions
are excluded from the module — tagging yourself isn't news.

Verified: module absent for an untagged account; visible with the
tagger named and linked for one that has tags; dismiss hides it; hard
reload keeps it hidden with the watermark intact. Browse link renders
logged-out. /tags/mine redirects to the Tags tab when signed in. Four
tabs on one line at 375px.

Known minor gap, pre-existing: RequireAuth builds `next` from pathname
only, so a logged-out visitor to /tags/mine lands on Settings/Profile
rather than Settings/Tags. Not worth touching auth redirect logic for.

145 tests, npm run check, and build clean.
